Make Bitmask a outer class
The functionality of Bitmask is very much independent of the class Tree.
It should be an independent class.
